requirements.txt
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
numpy==1.21.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2
scikit-learn==0.24.2

